Dev user

+++ services/auth-service/internal/domain/account/repository.go
@@ -5,7 +5,7 @@ import (
 )
 
 type Repository interface {
-	FindByID(ctx context.Context, id uint) (*Account, error)
+	FindByID(ctx context.Context, id int64) (*Account, error)
 	FindByUsername(ctx context.Context, username string) (*Account, error)
 	Save(ctx context.Context, account *Account) error
 }
