+++ services/auth-service/internal/domain/role/repository.go
@@ -3,6 +3,6 @@ package role
 import "context"
 
 type Repository interface {
-	FindByUserID(ctx context.Context, userID uint) ([]*Role, error)
-	GrantByUserID(ctx context.Context, userID uint, roleID []uint) error
+	FindByUserID(ctx context.Context, userID int64) ([]*Role, error)
+	GrantByUserID(ctx context.Context, userID int64, roleID []int64) error
 }
