+++ services/auth-service/internal/infrastructure/persistence/postgres/model/role_permission_po.go
@@ -3,7 +3,7 @@ package model
 import "time"
 
 type RolePermission struct {
-	RoleID       uint
-	PermissionID uint
+	RoleID       int64
+	PermissionID int64
 	GrantedAt    time.Time
 }
