+++ services/user-service/internal/domain/address/address.go
@@ -1,14 +1,14 @@
 package address
 
 type Address struct {
-	id        uint
-	userID    uint
+	id        int64
+	userID    int64
 	address   string
 	phone     string
 	recipient string
 }
 
-func New(id, userID uint, address, phone, recipient string) (*Address, error) {
+func New(id, userID int64, address, phone, recipient string) (*Address, error) {
 	if userID == 0 || address == "" || phone == "" || recipient == "" {
 		return nil, ErrInvalidAddress
 	}
